Handle durations and clean up template files

+++ generator/generator.go
@@ -57,14 +57,22 @@ func (g *Generator) LoadSchema(path string) error {
 	return Validate(g.doc)
 }
 
-func (g *Generator) RenderTemplate(name string, outPath string) error {
+func (g *Generator) RenderTemplate(name string, outPath string, clean bool) error {
 	if info, err := os.Stat(outPath); err != nil {
-		if !errors.Is(err, os.ErrNotExist) {
+		if errors.Is(err, os.ErrNotExist) {
+			err = os.Mkdir(outPath, os.ModePerm)
+			if err != nil {
+				return err
+			}
+		} else {
 			return err
 		}
-	} else if info.IsDir() {
-		log.Warn().Msgf("Removing existing output folder")
+	} else if info.IsDir() && clean {
 		os.RemoveAll(outPath)
+		err = os.Mkdir(outPath, os.ModePerm)
+		if err != nil {
+			return err
+		}
 	}
 
 	manifest, err := g.GetManifest(name)
@@ -80,11 +88,6 @@ func (g *Generator) RenderTemplate(name string, outPath string) error {
 	}
 
 	log.Debug().Msgf("Defined: %s", strings.TrimPrefix(template.DefinedTemplates(), "; defined templates are: "))
-
-	err = os.Mkdir(outPath, os.ModePerm)
-	if err != nil {
-		return err
-	}
 
 	for _, target := range manifest.Render {
 		log.Info().Msgf("Rendering target \"%s\"", target.Path)
@@ -102,7 +105,7 @@ func (g *Generator) RenderTemplate(name string, outPath string) error {
 
 		case "root":
 			filePath := path.Join(outPath, target.Path)
-			f, err := os.OpenFile(filePath, os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0600)
+			f, err := os.OpenFile(filePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0600)
 			if err != nil {
 				return err
 			}
@@ -120,7 +123,7 @@ func (g *Generator) RenderTemplate(name string, outPath string) error {
 			for _, tag := range g.doc.Tags {
 				targetPath := strings.Replace(target.Path, "*", tag.Name, 1)
 				filePath := path.Join(outPath, targetPath)
-				f, err := os.OpenFile(filePath, os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0600)
+				f, err := os.OpenFile(filePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0600)
 				if err != nil {
 					return err
 				}
